+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - New: add danger zone #9, not functional yet
 - Opt: add `loading="lazy"` to the `ImageUrl` renderer
 - Break: services will exchange messages instead of readings
+- Opt: use `multiqueue2`
 
 ## `0.15.0`